+++ packages/dapp-kit-next/packages/dapp-kit-core/src/web/internal/connected-account-menu-item.ts
@@ -4,13 +4,10 @@
 
 import { when } from 'lit/directives/when.js';
 import { html, LitElement } from 'lit';
-import { formatAddress } from '@socialproof/myso/utils';
 import { property, state } from 'lit/decorators.js';
 import type { UiWalletAccount } from '@wallet-standard/ui';
 import { copyIcon } from './icons/copy-icon.js';
-import { Task } from '@lit/task';
 import type { DAppKitCompatibleClient } from '../../core/types.js';
-import { resolveNameServiceName } from '../../utils/name.js';
 import { circleCheckIcon } from './icons/circle-check-icon.js';
 
 export type AccountSelectedEvent = CustomEvent<{ account: UiWalletAccount }>;
@@ -31,11 +28,6 @@ export class AccountMenuItem extends LitElement {
 
 	@state()
 	private _wasCopySuccessful = false;
-
-	#resolveNameTask = new Task(this, {
-		args: () => [this.client, this.account.address],
-		task: async ([client, address]) => resolveNameServiceName(client, address),
-	});
 
 	override connectedCallback() {
 		super.connectedCallback();
@@ -62,11 +54,6 @@ export class AccountMenuItem extends LitElement {
 				/>
 				<label class="content" for=${this.account.address}>
 					${when(this.account.icon, (icon) => html`<img src=${icon} alt="" />`)}
-					${this.#resolveNameTask.render({
-						pending: this.#renderAccountInfo,
-						complete: this.#renderAccountInfo,
-						error: () => this.#renderAccountInfo(),
-					})}
 				</label>
 				<button
 					class="copy-address-button"
@@ -93,17 +80,6 @@ export class AccountMenuItem extends LitElement {
 			// Do nothing here
 		}
 	}
-
-	#renderAccountInfo = (name?: string | null) => {
-		const { address, label } = this.account;
-		const formattedAddress = formatAddress(address);
-		const title = name || label;
-
-		return html`<div class="account-info">
-			<div class="account-title">${title || formattedAddress}</div>
-			${when(title, () => html`<div class="account-subtitle">${formattedAddress}</div>`)}
-		</div>`;
-	};
 
 	#accountSelected() {
 		this.dispatchEvent(