+++ packages/file-storage/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @socialproof/file-storage
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @socialproof/myso@0.0.4
+
 ## 0.0.3
 
 ### Patch Changes