+++ packages/walletconnect-wallet/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @socialproof/walletconnect-wallet
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @socialproof/myso@0.0.4
+  - @socialproof/wallet-standard@0.0.2
+
 ## 0.0.3
 
 ### Patch Changes